[abt-lib/AbtUtil] support not-necessarily-generative names in subst
This functor relied upon VARIABLE.clone generating a *fresh* variable;
however, I wish to leave that up to the semantics of the individual
VARIABLE implementation. In order for AbtUtil to be correct with respect
to these open-ended semantics, then, I just fixed it to check the
equality of the variable binding with the substituting variable, and if
it is equal, than do not bother substituting any further. (Since any
further occurrences of that name are "shadows".)