Beginning of intersection type; fix some issues with squash types
It was not possible to define intersections in their full generality in
terms of Squash. In MetaPRL, they have a USELESS intro rule for isect,
which requires you to provide in advance the witness. This is such a
horrible idea, and destroys refinement proof.

As such, I have reverted to Nuprl's approach, and had to add a notion of
"hidden" or "visible" hypothesis. It doesn't seem to have added too much
noise, thankfully.

I have also replaced the Squash elim rule with a stronger one (which is
still true).

(BTW, the Library is using still Context as its store, but this doesn't
make much sense now that Visibility information is present. Or does it?)