Add TELESCOPE signature & implementation
The purpose of this structure is to have both fast lookup by key, and
fast insertion, and to be traversable from either end quickly *in
order*. Space-usage may be not-so-good (I have not profiled it), but it
seems like at least for the things I list above, performance should be a
lot better than the two things that are typically done:

1. One option (which is what is usually done) is to use a linked list in
order to maintain order. This gives good traversal in one order, good
insertion at one end, bad insertion everywhere else, and bad lookup.

2. Another option (which is also frequently done) is to use a hashtable
or dictionary. Here, you get good insertion and lookup, but traversing
in order is quite poor.

3. Another option might be to use some kind of finger-tree like
structure, but it seemed like this would not be so good for looking up
by name. I toyed with doing something along these lines, whilst
simultaneously keeping a book of zipper-paths into the structure keyed
against labels, but this became super gnarly.