Slightly better behavior for level inference
Don't try and infer the level of a custom type operator; will revisit
when we add a version of this tactic that takes the state of the
development into account.

This "resolves" the issue noted by @darinmorrison
[here](https://github.com/jonsterling/JonPRL/commit/6f5c4b79e23fa1d6397a3f92cca8421e68f4cb17#commitcomment-11657785).

With this change, sometimes you have to pass a level annotation to a tactic.