Correct the ImageElim rule & derivation (details below)
(@vrahli --- In case you want to double check that I did not mess it up,
I've summarized my changes below.)

In Nuprl, full derivations are "ephemeral" (they exist in time, but not
as objects). On the other hand, JonPRL produces full derivations, which
then have their computational content "wringed out" of them through
extraction. This is mainly because I want to have the option in the
future to do interesting proof-theoretic things with JonPRL (such as
check its derivations in Coq, etc.).

As such, I am trying to make sure that the language of
derivations/proofs is a sensible one, which is why in any case that a
premise to a rule introduces a variable into the context, it must be
*bound* in the derivation. In the case of the ImageElim rule, we need to
increase the valence of its premise to bind the variable <w>.

The `extract` function takes a derivation and returns an extract term;
therefore, any subderivations must be passed through `extract` as well,
otherwise the extract will have derivations in it rather than
computational term.

In the case of ImageElim, this rule is only constructive in case the
motive of elimination is proof irrelevant (like an equality type)---so
it may seem like a waste of time to properly compute its extract in the
way I have done (see extract.fun), but it seems important to maintain
compositionality of extraction.